chore: Refactored ts to appropriate directories
