+++ index.html
@@ -30,10 +30,6 @@
   <!-- <script src="https://kit.fontawesome.com/33c5cad846.js" crossorigin="anonymous"></script> -->
   <!-- Advanced Animation Components -->
   <script src="js/ui-components/animated-x-scrollable.js"></script>
-  <!-- ./ts -->
-  <!-- <script type="module" src="ts/page-management.js"></script>
-  <script type="module" src="ts/page.js"></script>
-  <script type="module" src="ts/cubic-x-axis-transition.js"></script> -->
 </head>
 
 <body class="dynamic-bg">
@@ -324,10 +320,10 @@
   </div>
   
 
-  <script  type="module">
-    import {Page} from "/ts/page.js";
-    import {PageManagement} from "/ts/page-management.js";
-    import {CubicXAxisTransition} from "/ts/cubic-x-axis-transition.js";
+  <script type="module">
+    import {Page} from "/ts/engine/page.js";
+    import {PageManagement} from "/ts/engine/page-management.js";
+    import {CubicXAxisTransition} from "/ts/engine/cubic-x-axis-transition.js";
     var page1 = new Page("page-1", new CubicXAxisTransition());
     var page2 = new Page("page-2", new CubicXAxisTransition());
     var pageManagement = new PageManagement("page-management-container", [page1, page2]);
