+++ ts/engine/transition.js
